server/requirements.txt
alembic==0.9.3
aniso8601==9.0.1
bidict==0.21.2
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==6.7
cryptography==3.4.7
dnspython==1.16.0
dominate==2.3.1
eventlet==0.31.0
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.3.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
greenlet==1.1.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
Pillow==8.2.0
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.6.1
python-editor==1.0.3
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
requests==2.25.1
six==1.10.0
SQLAlchemy==1.3.13
tencentcloud-sdk-python==3.0.393
typing-extensions==3.10.0.0
urllib3==1.26.4
visitor==0.1.3
Werkzeug==0.12.2
wincertstore==0.2
WTForms==2.1
zipp==3.4.1
